Fix an issue into EffectParser with extended pass when properties already existed.
Details:
Let's say that we have a BaseEffect and an ExtendedEffect that extends the pass of the BaseEffect. If ExtendedEffect had declared some properties like attributes into its global scope, when we will load properties values of BaseEffect because of "extends" keyword, the pass attribute provider already contains the values present in ExtendedEffect global scope, so we need to merge both providers to keep the global scope values.